+++ packages/sdks/web-component/src/lib/types.ts
@@ -39,10 +39,13 @@ export enum Direction {
   forward = 'forward',
 }
 
-export interface LastAuthState {
+export type LastAuthState = NonNullable<
+  NextFnReturnPromiseValue['data']['lastAuth']
+> & {
   loginId?: string;
   name?: string;
-}
+  lastUsedPerScreen?: Record<string, string>;
+};
 
 export interface ScreenState {
   errorText?: string;
@@ -71,6 +74,7 @@ export interface ScreenState {
 export type SSOQueryParams = {
   oidcIdpStateId?: string;
   samlIdpStateId?: string;
+  wsfedIdpStateId?: string;
   samlIdpUsername?: string;
   descopeIdpInitiated?: boolean;
   ssoAppId?: string;
@@ -122,6 +126,9 @@ export type FlowState = {
   samlIdpResponseUrl: string;
   samlIdpResponseSamlResponse: string;
   samlIdpResponseRelayState: string;
+  wsFedIdpResponseUrl: string;
+  wsFedIdpResponseWresult: string;
+  wsFedIdpResponseWctx: string;
   nativeResponseType: string;
   nativePayload: Record<string, any>;
   reqTimestamp: number;