+++ packages/sdks/web-js-sdk/CHANGELOG.md
@@ -2,6 +2,48 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.48.4](https://github.com/descope/descope-js/compare/web-js-sdk-1.48.3...web-js-sdk-1.48.4) (2026-05-14)
+
+
+### Bug Fixes
+
+* **web-js-sdk:** call afterRequest hook in OIDC loginWithRedirect before navigation ([#1385](https://github.com/descope/descope-js/issues/1385)) ([dcb983c](https://github.com/descope/descope-js/commit/dcb983cfd5a690da9a6e63644802a59d8ef850b0))
+
+## [1.48.3](https://github.com/descope/descope-js/compare/web-js-sdk-1.48.2...web-js-sdk-1.48.3) (2026-05-11)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.61.0`
+## [1.48.2](https://github.com/descope/descope-js/compare/web-js-sdk-1.48.1...web-js-sdk-1.48.2) (2026-04-29)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.60.0`
+## [1.48.1](https://github.com/descope/descope-js/compare/web-js-sdk-1.48.0...web-js-sdk-1.48.1) (2026-04-21)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.59.1`
+## [1.48.0](https://github.com/descope/descope-js/compare/web-js-sdk-1.47.2...web-js-sdk-1.48.0) (2026-04-20)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.59.0`
+
+### Features
+
+* change flow api version to v2 ([#1323](https://github.com/descope/descope-js/issues/1323)) ([756096e](https://github.com/descope/descope-js/commit/756096ecff1678686a2ad77f02428fb1bcf49535))
+
+## [1.47.2](https://github.com/descope/descope-js/compare/web-js-sdk-1.47.1...web-js-sdk-1.47.2) (2026-04-20)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.58.2`
+
+### Bug Fixes
+
+* **wc:** add wsfedIdpStateId URL param support for WS-Fed IDP flow RELEASE ([#1367](https://github.com/descope/descope-js/issues/1367)) ([b146bec](https://github.com/descope/descope-js/commit/b146bec3670ae33974331731929482e884d23c85)), closes [descope/etc#15054](https://github.com/descope/etc/issues/15054) [descope/descope#532](https://github.com/descope/descope/issues/532)
+
 ## [1.47.1](https://github.com/descope/descope-js/compare/web-js-sdk-1.47.0...web-js-sdk-1.47.1) (2026-04-13)
 
 ### Dependency Updates